Reorganise according to PSR4 etc.
