+++ .travis.yml
@@ -1,5 +1,4 @@
 sudo: false
-dist: trusty
 language: php
 php:
 - '5.5'
