+++ CODE_OF_CONDUCT.md
@@ -1,0 +1,1 @@
+The development of this software is covered by a [Code of Conduct](https://www.mediawiki.org/wiki/Code_of_Conduct).
