+++ composer.json
@@ -23,8 +23,8 @@
 		"mediawiki/mediawiki-codesniffer": "^14.0"
 	},
 	"autoload": {
-		"psr-0": {
-			"IaUpload\\": "src/"
+		"psr-4": {
+			"Wikisource\\IaUpload\\": "src/"
 		}
 	},
 	"scripts": {
