+++ src/OAuth/Token/Token.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace IaUpload\OAuth\Token;
+namespace Wikisource\IaUpload\OAuth\Token;
 
 use JsonSerializable;
 use Serializable;
